Criteria for Retirement Location

A dataset containing criteria for relocation in retirement.

data("retirementLoc")

Format

A data frame with 3146 rows and 21 variables:

fips

a five digit unique identifier of US counties

lat

latitude of county center in degrees

lon

longitude of county center in degrees

state

state name

county

county name

pop_2020

population estimate in 2020

pct_pop_change

percent population change from 2010 to 2020

cbsa_desig

a classification based on population density

rucc_2013

a classification of 1 through 9 on a rural-urban continuum

partisan_lean

percent voting Democrat in 2020 presidential race

med_hh_inc_2019

median household income for 2019

pct_bachelor

percent of county residents holding a bacherlor degree or higher

broadband_2017

percent households with broadband access

life_exp

life expectancy in years

violent_crime_rate

violent crimes per 100,000

average_daily_pm2_5

Average daily amount of fine particulate matter in micrograms per cubic meter

prim_care_dr_rate

primary care physicians per 100,000 population

avg_annual_temp

average annual temperature in fahrenheit

median_home_price

median home price

yoy_price_chg_pct

year-over-year percentage increase (-decrease) in home price

years_to_payoff

quotient of household income divided by median home price

Details

The dataset includes the 50 US states and Washington D.C. disaggregated at the county level.

Examples

#load
data("retirementData")

#> Warning: data set ‘retirementData’ not found

sum(retirementData::retirementLoc$pop_2020)

#> [1] NA